Refine ocr dygraph code, add infer module (#4182)
* refine code, add infer module

* update readme
