+++ dygraph/ocr_recognition/README.md
@@ -25,11 +25,27 @@ ocr任务是识别图片单行的字母信息，在动态图下使用了带atten
 在GPU单卡上训练ocr recognition:
 
 ```
-env CUDA_VISIBLE_DEVICES=0 python train.py
+CUDA_VISIBLE_DEVICES=0 python train.py
 ```
 
 这里`CUDA_VISIBLE_DEVICES=0`表示是执行在0号设备卡上，请根据自身情况修改这个参数。
 
-## 效果
+## 测试ocr recognition
 
-在test测试集合上，最好的效果为82.0%
+
+```
+CUDA_VISIBLE_DEVICES=0 python eval.py --pretrained_model your_trained_model_path
+```
+
+## 预测
+
+
+```
+CUDA_VISIBLE_DEVICES=0 python -u infer.py --pretrained_model your_trained_model_path --image_path your_img_path
+```
+
+## 预训练模型
+
+|模型| 准确率|
+|- |:-: |
+|[ocr_attention_params](https://paddle-ocr-models.bj.bcebos.com/ocr_attention_dygraph.tar) | 82.46%|
