--- dygraph/ocr_recognition/debug.sh
@@ -1,4 +1,0 @@
-
-export CUDA_VISIBLE_DEVICES=0
-
-python train.py
